refactor csv report
extracted csv report into separate module. removed all maven related stuff rrom csv report
