+++ pom.xml
@@ -58,7 +58,7 @@
         <javac.version>9+181-r4173-1</javac.version>
 
         <!--Compiler plugins-->
-        <errorprone.version>2.10.0</errorprone.version>
+        <errorprone.version>2.20.0</errorprone.version>
         <lombok.version>1.18.28</lombok.version>
 
         <spotbugs.maven.plugin.version>4.0.0</spotbugs.maven.plugin.version>
