+++ report/src/test/java/org/hjug/refactorfirst/report/HtmlReportTest.java
@@ -1,10 +1,9 @@
 package org.hjug.refactorfirst.report;
 
-import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.Locale;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import org.junit.jupiter.api.Test;
 
 public class HtmlReportTest {
